+++ 1.21.4/Common/src/main/java/fuzs/puzzleslib/api/client/data/v2/AbstractModelProvider.java
@@ -24,7 +24,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.function.Predicate;
 
 public abstract class AbstractModelProvider implements DataProvider {
-    private final String modId;
+    protected final String modId;
     private final PackOutput.PathProvider blockStatePathProvider;
     private final PackOutput.PathProvider itemInfoPathProvider;
     private final PackOutput.PathProvider modelPathProvider;