+++ 1.21.4/Common/src/main/java/fuzs/puzzleslib/api/core/v1/context/BlockInteractionsContext.java
@@ -8,6 +8,7 @@ import org.jetbrains.annotations.Nullable;
 /**
  * Register various block transformations triggered by right-clicking with certain vanilla tools.
  */
+@Deprecated(forRemoval = true)
 public interface BlockInteractionsContext {
 
     /**