+++ 1.21.4/Common/src/main/java/fuzs/puzzleslib/api/core/v1/context/FuelValuesContext.java
@@ -8,6 +8,7 @@ import net.minecraft.world.level.ItemLike;
 /**
  * Register items as furnace fuel.
  */
+@Deprecated(forRemoval = true)
 public interface FuelValuesContext {
 
     /**