+++ 1.21.4/Fabric/src/main/java/fuzs/puzzleslib/fabric/impl/core/FabricModConstructor.java
@@ -15,6 +15,7 @@ public final class FabricModConstructor {
         constructor.onEntityAttributeCreation(new EntityAttributesCreateContextFabricImpl());
         constructor.onEntityAttributeModification(new EntityAttributesModifyContextFabricImpl());
         constructor.onRegisterSpawnPlacements(new SpawnPlacementsContextFabricImpl());
+        constructor.onRegisterGameplayContent(new GameplayContentContextFabricImpl());
         constructor.onRegisterFlammableBlocks(new FlammableBlocksContextFabricImpl());
         constructor.onRegisterFuelValues(new FuelValuesContextFabricImpl());
         constructor.onRegisterCompostableBlocks(new CompostableBlocksContextFabricImpl());