+++ 1.21.4/NeoForge/src/main/java/fuzs/puzzleslib/neoforge/impl/core/NeoForgeModConstructor.java
@@ -24,6 +24,7 @@ public final class NeoForgeModConstructor {
             constructor.onConstructMod();
             // these need to run immediately, as they register content for data generation,
             // which cannot be added during common setup, as it does not run during data generation
+            constructor.onRegisterGameplayContent(new GameplayContentContextNeoForgeImpl(modId));
             constructor.onRegisterCompostableBlocks(new CompostableBlocksContextNeoForgeImpl(modId));
             constructor.onRegisterFuelValues(new FuelValuesContextNeoForgeImpl(modId));
             constructor.onRegisterBiomeModifications(new BiomeModificationsContextNeoForgeImpl(modId, eventBus));