+++ 1.21.4/NeoForge/src/main/java/fuzs/puzzleslib/neoforge/impl/core/context/FlammableBlocksContextNeoForgeImpl.java
@@ -2,9 +2,9 @@ package fuzs.puzzleslib.neoforge.impl.core.context;
 
 import com.google.common.base.Preconditions;
 import fuzs.puzzleslib.api.core.v1.context.FlammableBlocksContext;
-import fuzs.puzzleslib.neoforge.mixin.accessor.FireBlockNeoForgeAccessor;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.FireBlock;
 
 import java.util.Objects;
 
@@ -18,7 +18,7 @@ public final class FlammableBlocksContextNeoForgeImpl implements FlammableBlocks
         Preconditions.checkState(blocks.length > 0, "blocks is empty");
         for (Block block : blocks) {
             Objects.requireNonNull(block, "block is null");
-            ((FireBlockNeoForgeAccessor) Blocks.FIRE).puzzleslib$setFlammable(block, encouragement, flammability);
+            ((FireBlock) Blocks.FIRE).setFlammable(block, encouragement, flammability);
         }
     }
 }