+++ 1.21.1/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.1.31-1.21.1] - 2025-03-13
+### Added
+- Add `ModConstructor::onRegisterGameplayContent`
+- Add `ModelLocationHelper` & `ModelTemplateHelper`
+- Add `ResourceKeyHelper`
+- Add `AbstractDatapackRegistriesProvider`
+
 ## [v21.1.30-1.21.1] - 2025-03-11
 ### Fixed
 - Fix rare start-up crash when registering generated config translations on NeoForge